Components: Rename folders to camelCase
closes #315
