+++ packages/components/src/illustration/Illustration.js
