+++ packages/components/src/illustration/IllustrationTypes.js
