+++ packages/components/src/illustration/assets/mobile/Boarding.png
