+++ packages/components/src/illustration/assets/mobile/Boarding@2x.png
