+++ packages/components/src/illustration/assets/mobile/Boarding@3x.png
