+++ packages/components/src/illustration/assets/mobile/Error.png
