+++ packages/components/src/illustration/assets/mobile/Error@2x.png
