+++ packages/components/src/illustration/assets/mobile/Error@3x.png
