+++ packages/components/src/illustration/assets/mobile/NoResults.png
