+++ packages/components/src/illustration/assets/mobile/NoResults@2x.png
