+++ packages/components/src/illustration/assets/mobile/NoResults@3x.png
