+++ packages/components/src/placeCard/__tests__/index.test.js
@@ -5,7 +5,7 @@ import { render, fireEvent } from 'react-native-testing-library';
 import { Touchable } from '@kiwicom/universal-components';
 
 import PlaceCard from '../PlaceCard';
-import MissingImage from '../Image/MissingImage';
+import MissingImage from '../image/MissingImage';
 
 describe('PlaceCard', () => {
   const onPress = jest.fn();
