+++ packages/components/src/placeCard/assets/black-to-alpha-vertical.png
