+++ packages/components/src/placeCard/image/ImageTypes.js
