+++ packages/components/src/searchParamsSummary/__tests__/index.test.js
