+++ packages/components/src/searchParamsSummary/assets/alpha-to-white-horizontal.png
