+++ packages/components/src/searchParamsSummary/index.js
