+++ packages/components/src/shimmer/Shimmer.native.js
